fixed april tags camera issue.
