+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AprilTagParking2939.java
@@ -22,6 +22,8 @@
 package org.firstinspires.ftc.teamcode.Autonomous;
 
 import org.firstinspires.ftc.teamcode.RobotConstants;
+
+import static org.firstinspires.ftc.robotcore.external.BlocksOpModeCompanion.hardwareMap;
 import static org.firstinspires.ftc.teamcode.RobotConstants.Commands.DRIVE.FORWARD;
 import static org.firstinspires.ftc.teamcode.RobotConstants.Commands.DRIVE.BACKWARD;
 import static org.firstinspires.ftc.teamcode.RobotConstants.Commands.DRIVE.LEFT_STRAFE;
@@ -56,10 +58,17 @@ public class AprilTagParking2939 extends LinearOpMode {
     static final double TICKS_PER_INCH = 50;
     static final double TICKS_PER_DEGREE = 50;
 
+    double fx = 578.272;
+    double fy = 578.272;
+    double cx = 402.145;
+    double cy = 221.506;
+
+    double tagsize = 0.166;
+
     @Override
     public void runOpMode() throws InterruptedException {
         // Initialize the robot
-        RobotCentricMechanumDrive drive = new RobotCentricMechanumDrive(hardwareMap, DcMotorSimple.Direction.REVERSE);
+        RobotCentricMechanumDrive drive = new RobotCentricMechanumDrive(hardwareMap, DcMotorSimple.Direction.FORWARD);
         drive.setTicks(TICKS_PER_INCH, TICKS_PER_INCH);
 
         // Initialize Operator Members and Motors
@@ -68,18 +77,35 @@ public class AprilTagParking2939 extends LinearOpMode {
 
         // Initialize camera
         telemetry.setMsTransmissionInterval(50);
-        AprilTagsCamera camera = new AprilTagsCamera();
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        OpenCvCamera camera = OpenCvCameraFactory.getInstance().createWebcam(hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
+        AprilTagDetectionPipeline aprilTagDetectionPipeline = new AprilTagDetectionPipeline(tagsize, fx, fy, cx, cy);
+
+        camera.setPipeline(aprilTagDetectionPipeline);
+        camera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
+            @Override
+            public void onOpened() {
+                camera.startStreaming(800, 448, OpenCvCameraRotation.UPRIGHT);
+            }
+
+            @Override
+            public void onError(int errorCode) {
+
+            }
+        });
+
+        AprilTagsCamera aprilTagsCamera = new AprilTagsCamera(aprilTagDetectionPipeline);
 
         AprilTagDetection tagOfInterest = null;
         while (!isStarted() && !isStopRequested()) {
-            tagOfInterest = camera.detectTag(telemetry);
+            tagOfInterest = aprilTagsCamera.detectTag(telemetry);
         }
 
         waitForStart();
-        camera.updateTelemetry(telemetry);
+        aprilTagsCamera.updateTelemetry(telemetry);
 
         // Retrieve the position
-        int position = camera.getPosition(tagOfInterest);
+        int position = aprilTagsCamera.getPosition(tagOfInterest);
         telemetry.addLine("We are heading towards position #" + position + "!");
         switch (position) {
             case 1:
