+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/AprilTagsCamera.java
@@ -13,7 +13,6 @@ import org.openftc.easyopencv.OpenCvCameraRotation;
 import java.util.ArrayList;
 
 public class AprilTagsCamera {
-    OpenCvCamera camera;
     AprilTagDetectionPipeline aprilTagDetectionPipeline;
 
     static final double FEET_PER_METER = 3.28084;
@@ -32,23 +31,8 @@ public class AprilTagsCamera {
 
     AprilTagDetection tagOfInterest = null;
 
-    public AprilTagsCamera() {
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-        camera = OpenCvCameraFactory.getInstance().createWebcam(hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
-        aprilTagDetectionPipeline = new AprilTagDetectionPipeline(tagsize, fx, fy, cx, cy);
-
-        camera.setPipeline(aprilTagDetectionPipeline);
-        camera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
-            @Override
-            public void onOpened() {
-                camera.startStreaming(800, 448, OpenCvCameraRotation.UPRIGHT);
-            }
-
-            @Override
-            public void onError(int errorCode) {
-
-            }
-        });
+    public AprilTagsCamera(AprilTagDetectionPipeline aprilTagDetectionPipeline) {
+        this.aprilTagDetectionPipeline = aprilTagDetectionPipeline;
     }
 
     public void tagToTelemetry(AprilTagDetection detection, Telemetry telemetry) {
